requirements.txt
flask
flask-cors
pyjwt
protobuf